carbonsense-backend/requirements.txt
# Django
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
gunicorn>=21.2,<22.0

# Database
psycopg2-binary>=2.9,<3.0  # PostgreSQL adapter (optional, for production)
dj-database-url>=2.1,<3.0

# CORS handling
django-cors-headers>=4.3,<4.4

# Environment variables
python-dotenv>=1.0,<1.1

# Development
django-extensions>=3.2,<3.3  # Useful development utilities

# RAG / AI
chromadb>=0.4.22,<0.5
sentence-transformers>=2.2,<3.0
groq>=0.12,<1.0
google-generativeai>=0.8,<1.0
openai>=1.40,<2.0
tavily-python>=0.5,<1.0

# Document processing
PyPDF2>=3.0,<4.0
requests>=2.31,<3.0

# Web scraping & search
feedparser>=6.0,<7.0
beautifulsoup4>=4.12,<5.0
ddgs>=6.0

# Token counting
tiktoken>=0.5,<1.0

# Testing (optional)
pytest>=7.4,<8.0
pytest-django>=4.7,<4.8
